Get rid of TaskScopeServices
The `TaskScopeServices` service was created for each task, in hope we could lazily initialize some dependents. Typically,
task inputs and outputs were created using this. However, in practice, they are always eagerly initialized during the
task creation. This leads to both increased memory usage (because each task has a companion service) and performance
penalty (because injecting services as well as calling `get` on them is expensive).

This commit replaces the use of `TaskScopeServices` with eagerly created instances.

+review REVIEW-5921