Include sources in binary jars
This change is necessary in order to support quick documentation and
source navigation for Kotlin-based build scripts in IDEA. It is a
temporary workaround and will be reverted as soon as a more complete
solution is available in the Kotlin IDEA plugin. Overall, the inclusion
of source files adds approximately 5 MB to the `-bin` distribution zip.

This commit was originally introduced via merge in
685785fa9cda3d186fd958097d4e5937fa1545a0 and was subsequently reverted
in 377737a7c0649f081cbd298da027d3b804897aea due to a Javadoc-related CI
build failures it caused. This commit is being applied again now in
order to attempt to produce that same Javadoc-related CI failure once
again, as it has proven difficult to reproduce locally.