Attempt to fix flaky `ApplyPluginIntegSpec`
This test declares and executes a gradle build, 
that compiles and runs a unit test,
that in turn uses `ProjectBuilder` to mock a Gradle project,
that applies the 'groovy' plugin.
Turtles all the way down.

This fix ensures that the inner test specifies the Gradle project directory
for the inner-inner Gradle project to use, so that this directory can be
cleaned up correctly by the owning integration test. (Without configuration,
the `ProjectBuilder` will use a directory under 'tmp' using `deleteOnExit`)

The `ProjectBuilder` may initialize `NativeServices` in it's project directory.
I suspect this might be causing the DaemonExecuter to die, because unless
NativeServices has already been initialized in the Daemon process, it will
be initialized in the to-be-deleted directory under `tmp`.