GRADLE-2263 allow daemon instances to be shutdown. (#643)
This is a proposed fix for the issue described in:

https://issues.gradle.org/browse/GRADLE-2263

In GRADLE-2263, a workaround is described that added to a projects  build.gradle
script. This commit is a simple reordering of two statements that are described in that Jira. This
reordering of code effectively implements, for all users,  the workaround described in
that Jira.

This commit simply reorders these two operations so that a monitor port is created prior to performing
the exit if daemon check.