Exclude gradle-script-kotlin from core lib version checks
To date, all "core libs" have been co-versioned with the distribution.
gradle-script-kotlin is different. It is included in the distribution
with its own independent version. At the time of this writing, that
version is 1.0.0-M1a. This means that gradle-script-kotlin should not be
subject to the same version checking that other locally-managed core
libs go through.

This commit is a quick fix, simply excluding gradle-script-kotlin from
the version check by name. A more complete approach should be considered
in the future, particularly as additional externally-managed libs are
introduced to the distribution in this fashion.