Resolve from /libs-releases vs /libs-releases-local
Prior to this change, the affected build scripts resolved dependencies
from two repositories: /libs-releases-local and mavenCentral(), and in
that order. /libs-releases-local is, in Artifactory terms, a "local
repository", and it contains only the things we upload to it, e.g. our
gradle-core, gradle-logging, and other gradle-* modules. This means that
the vast majority of dependency resolution requests will fail to find
what they're looking for in /libs-release-local and will then fall back
to mavenCentral(). This is inefficient and slows down builds, because it
requires two client-side HTTP requests for almost every dependency.

This commit favors the use of the /libs-release "virtual repository",
which includes all the content from /libs-release-local, but also
proxies content from maven central (actually, it proxies content from
JFrog's JCenter, but JCenter itself is a proxy for Maven Central).

This means that just one repository declaration can be used instead of
two, and that the client build need only ever issue one HTTP request for
each dependency.

Note that one exception to the explanation above is
buildSrc/build.gradle, which is being changed from /libs to
/libs-release. This same change was recently made to the project's root
build.gradle as well, and the rationale for doing so can be found at
bf9ae865daacc4757615359586588af5b8c15909.