Resolve releases from /libs-releases
Prior to this commit, the /libs-milestones virtual repository had been
configured to include /libs-releases, just as libs-snapshots had been
configured to include /libs-milestones. This "chaining" of virtual
repositories had the positive effect of providing builds with "one stop
shopping", i.e. the need to declare only one repository, but it also
came with the negative effect of effectively associating release
artifacts with non-release repositories on the client side. This may be
problematic for Gradle's own artifact caching, and is in any case
confusing to observe from the Gradle command line. One wonders, for
example, while a build is resolving dependencies "why is jackson-core
2.3.2 being downloaded from repo.gradle.org/gradle/libs-milestones?"

These repositories have been updated on the Artifactory side such that
they no longer do this "chaining". The resulting effect is that builds
must declare each of the repositories they wish to resolve from. If a
build depends only on GA release artifacts, resolve from /libs-releases
(which still proxies all remote release repositories, including
JCenter); if a build depends on milestone artifacts, resolve from
/libs-milestones as well; and likewise, snapshot artifacts should be
resolved from /libs-snapshots.

The performance impact on builds should be quite minimal, so long as the
repositories are declared in the correct order: /libs-releases, then
/libs-milestones, then /libs-snapshots, assuming this order corresponds
to where the bulk of the build's dependencies actually reside. Most
resolutions, then, will require only a single HTTP request against
/libs-releases; if the dependency is not found there, /libs-milestones
and /libs-snapshots will be checked (if indeed either or both are
declared).