Introduce `TaskInputs/TaskOutputs.configure()`
This allows delayed configuration of inputs and outputs using the fully
configured task. When using e.g. `TaskOutputs.files()` the plugin could
supply a closure that would be resolved right before the task is
executed. To deprecate `TaskOutputs.files()` there needs to be a way to
do the same by calling `TaskOutputs.file()`.

+review REVIEW-6010