Use a build time generated relocation list
This commit changes the implementation of shaded jar relocation by using a list of packages generated at build time instead of a hardcoded
list of packages (`relocated.txt`). This allows us to make sure that whatever dependency we add to Gradle is taken into consideration when
generating the shaded jar.

The generation of the list of packages is implemented as a Gradle task. However, we can't use it directly in our build files because the
version of Gradle we use does not have this task. Therefore, it is **temporarily** copied into the `buildSrc` directory. This must be
removed as soon as we upgrade the wrapper to a build that contains the task.

+review REVIEW-6039