Rework plural task output registration APIs
We are reverting back to using the plural annotations `@OutputFiles` and `@OutputDirectories`, but only allow them to be used on `Map` properties (otherwise show a deprecation warning). The `Map` must associate a unique String identifier to each output file or directory, respectively.

This allows the removal of the recently introduced `configure()` methods on `TaskInputs` and `TaskOutputs`, but requires the introduction of the `TaskOutputs.namedFiles(Map)` method (and a version that takes a `Callable` to produce the map).

+review REVIEW-6055