Make sure all `JavaCrossCompilationIntegrationTest` passes with latest Java 9
This commit adds Java 9 in the list of target Java versions for `JavaCrossCompilationIntegrationTest`. It introduces a temporary
workaround in `ClassFile` to make it aware of the Java 9 class format, waiting for ASM 6 to be available. It also removes the
temporary workaround on `JavaPluginConvention` for Java 9 classes.

Note that the Java versioning scheme changes from Java 9. Previous versions uses `1.1`, `1.2`, ... but Java 9 now uses `9`. The
change was introduced in recent builds of Java 9, so for this test to pass, one must make sure to use the latest Java 9 EA builds.

+review REVIEW-6054