Fix JUnitIntegrationTest on Java 9
For Java 9 we use `@` as command line parameter to launch
the Worker Process for test execution. By doing this we no longer
need to use a `SecurityManager`-hack which loads the classpath
on startup to work around Windows classpath length issues.
Moreover, on Java 9 the System Classloader is no longer an URLClassloader -
we cannot determine the list of Jars from this.

+review REVIEW-6061