Separate construction of composite-aware services from registering participants
The `CompositeContextBuilder` service allows participants to be registered
with the `CompositeBuildContext`, so that they are available to dependency
resolution in a composite build. By separating the construction of the
context (and builder) from the actual registration, we are now able to
add/register participants in different ways at different points of the
build lifecycle.