Create getters for constants for converted Classes
In Groovy, a public static constant is converted into a private
constant and public static getters. In order to not break any plugins
we now generate these getters at runtime. Since we only have String
constants only this is supported.
By this we do not break binary compatibility. As soon as the plugin is
recompiled the reference to the getter is not any more in the bytecode.

We plan to remove these mixings for 4.0