Create get getters for booleans for converted Classes
In Groovy, a public boolean field converted into a private
field and public get getters, is getters and setters. In order to not break any plugins
we now generate these get getters at runtime.
By this we do not break binary compatibility. As soon as the plugin is
recompiled the reference to the getter is not any more in the bytecode.

We plan to remove these mixings for 4.0

+review REVIEW-6102