Reverted new public API for task properties
When introducing new APIs return types of methods like `TaskInputs.files()` were narrowed to a subtype of the original return type, and this broke binary compatibility with Gradle 2.x. This commit adds interfaces `TaskInputsVersion2` and `TaskOutputsVersion2` that re-introduces the Gradle 2.x-compatible method signatures while keeping the new ones introduced in Gradle 3.x.

The new interfaces are made public in order not to have to refer to internal types from the public API. The new types are also deprecated, and are going to be removed in Gradle 4.0.

+review REVIEW-6125