Attempt to fix `ConfigureLogging` rule
The `ConfigureLogging` rule is broken. It replaces the configured `OutputEventListener`
for the SLF4J logging system. Then, after the tests completes it resets the SLF4J logging
back to it's default state.

The `InProcessGradleExecuter` attaches a listener so that it can capture any logging output
created during a test build invocation. This is done only once for the global logging
system. The `ConfigureLogging` rule doesn't reattach this listener when it completes
so any subsequent test will not have some missing log events when executing builds with `InProcessGradleExecuter`.

This change causes `ConfigureLogging` to reinstate the original `OutputEventListener` when resetting the logging.

Fix stupid mistake in logging experiment
(cherry picked from commit 8fd0c8b)