Fix for `EmptyStackException` processing task execution plan
This is a more complete fix for GRADLE-3127.

In certain complex task graphs containing shouldRunAfter cycles, an
EmptyStackException is thrown building the task execution plan. 

This change cleans up walked shouldRunAfter edges from the stack
in the case where a node is visited multiple times when building the
task execution plan.