Remove @Deprecated from Reporting#reports(Closure)
We do not have a strategy yet what to do with Closure taking methods.
Deprecating doesn't seem to be an option because then you see
deprecation warnings when writing `build.gradle` files.

For Gradle 3.0 we will not deprecate these methods and think
about a strategy for the future which we will implement later
in the 3.x release cycle

+review REVIEW-6142