Allow a composite build to declare the outputs provided by included build
By default, we configure each included build to determine which modules
should be substituted in the composite build. In many cases this automatic
detection of outputs is not sufficient, and this change allows a user to
explicitly declare the outputs provided by an included build.

The new DSL for declaring outputs is merely a placeholder to prove this
functionality. More work will be done to polish this interface.