Execute performance test scenarios on a fresh working copy
Reusing whatever state the last test left behind can make performance
seem better (because of preexisting caches) or worse (because of lots of output).
This makes the results dependent on the order in which the tests are executed.

It also prevented us from using incremental build for the project templates.

We now create a fresh copy of the template project for each test run,
fixing both of these problems at once.