Added methods to convert an immutable `ModuleComponentResolveMetadata` into a mutable `MutableModuleComponentResolveMetadata` and vice versa.
The implementation doesn't quite honour the contract yet, as each view implementation is actually mutable. The immutable view is immutable in the sense that it can't be changed but the implementation computes some values lazily and so the state is not thread safe.

This is an initial step to move all the consumers of `ModuleComponentResolveMetadata` to statically express whether they mutate the view or not. The goal is to have a strictly immutable implementation that can be shared across threads without requiring a defensive copy to be made.