Put all cross-build results into one database
We had one database for each type of cross-build result (e.g.
Gradle vs. Maven results) because there was no way of
telling which type a row in the database had. This obviously
does not scale when we want to create more kinds of comparisons.

Instead, the database now has a resultType field which is used to
discriminate between the different types of results.

As a next step, we still need to unify the databases
for cross-build and cross-version results.