Allow specifying the list of baselines from command line
This commit introduces the ability to set the list of baseline versions for performance tests through a command-line
flag (`--baselines`) or a system property (`org.gradle.performance.baselines`). The list must be specified as a
comma-separated list of versions or as a semicolon separated list of versions.

Two versions are handled particularly:

* `last` corresponds to the last release of Gradle
* `nightly` corresponds to the latest build of Gradle (`master` branch)

This commit also removes the "ad-hoc" mode for executing tests. The idea is to replace it with this flag, which
can be set to `nightly`. However, the "ad-hoc" mode skipped writing results to the database, in order to avoid
polluting the performance results DB with tests. If you want to do this, you need to set the `gradle.performance.db.url`
to a local, temporary database:

```
./gradlew cleanSmallOldJava smallOldJava cleanPerformanceTest performanceTest --scenarios 'clean Java build smallOldJava (daemon)' -x prepareSamples --baselines nightly -Porg.gradle.performance.db.url=jdbc:h2:./build/database
```